refactor: 메타데이터 어댑터·동기화 폴링 공용화 + 다운로드 유틸 추출 (#697)
- utils/metadataAdapters.js: MetadataPickerModal 의 KIND_ADAPTERS 와
  MetadataManagementBody 의 ADAPTERS 를 슈퍼셋으로 통합 (fetch 시그니처는
  picker 형태의 단일 객체 인자로 표준화, admin 전용 필드 포함)
- hooks/useMetadataSyncPolling.js: 2초 동기화 폴링 + 초기 상태 확인
  (checkInitialSyncStatus) 공용 hook — adapter/onCompleted 는 ref 로 고정해
  기존 양쪽의 eslint-disable 재발 없이 안정적 의존성 유지
- utils/download.js: downloadBlob/downloadFromUrl — 뷰어/카드 5곳의
  createObjectURL+<a download> 중복 제거 (toast 는 호출부 유지)

동작 보존 검증:
- jest 307, vite 빌드, docker-compose 재빌드 + e2e (기존 #381 2건 외 통과)
- 실 브라우저(Playwright): 작업판 실행 화면 → LoRA picker 오픈 → 공용 어댑터
  fetch 200 → 목록/빈 상태 렌더, 콘솔 에러 0 (admin 페이지는 #381 제약으로
  실행 불가 — 동일 공용 경로 사용)
